requirements.txt
requests
pandas
scikit-learn
streamlit
shap
python-dotenv
hopsworks
hopsworks[python]
azure-storage-blob
matplotlib
flake8
autopep8
black
ruff
pre-commit
